Be more robust against premature process crash.
In case a process crashes right after calling `attachApplication`, it is
possible that binder wasn't able to get the PID of the process while
delivering the transaction, resulting in a PID of 0. libprocessgroup's
killProcessGroup() requires a positive PID, so only call when we got a
valid PID.

Bug: 297833427
Test: N/A

+++ services/core/java/com/android/server/am/ActivityManagerService.java
@@ -4869,8 +4869,10 @@ public class ActivityManagerService extends IActivityManager.Stub
         } else {
             Slog.wtf(TAG, "Mismatched or missing ProcessRecord: " + app + ". Pid: " + pid
                     + ". Uid: " + uid);
-            killProcess(pid);
-            killProcessGroup(uid, pid);
+            if (pid > 0) {
+                killProcess(pid);
+                killProcessGroup(uid, pid);
+            }
             mProcessList.noteAppKill(pid, uid,
                     ApplicationExitInfo.REASON_INITIALIZATION_FAILURE,
                     ApplicationExitInfo.SUBREASON_UNKNOWN,
